delete render context
